Bugfix: Send correct parameters to CreateMover
The "E:LoadMovers()" call to the local function "CreateMover()"
was missing its "text" parameter (3rd parameter), which meant that
all subsequent function calls via LoadMovers submitted corrupted
function parameters.

However, the proper value for the "text" parameter IS being saved by
"E:CreateMover()", so we HAVE everything we needed to fix the bug. We
simply have to pass the t.text value in the call!